// frontend/src/app/core/api/models/ai-settings-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AiSettingsResponse {
  classifier_base_url?: (string | null);
  classifier_model: string;
  classifier_provider: string;
  id: string;
  responder_base_url?: (string | null);
  responder_model: string;
  responder_provider: string;
}